add a new file

+++ app/src/main/java/com/example/babistone_auth/model/fashio_model.kt
@@ -1,5 +1,5 @@
 package com.example.babistone_auth.model
 
-class fashio_model(val nom:String, val prix:Int,val image:Int) {
+class fashio_model(val nom:String, val prix:Double,val image:Int) {
 
 }
